pyproject.toml
[project]
name = "financialagent"
version = "0.1.0"
description = "A LangGraph-based agent for performing stock analysis."
readme = "README.md"
requires-python = ">=3.9" # Changed to a more common Python version
dependencies = [
    "pandas>=2.0.0",
    "yfinance>=0.2.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.0.30",
    "python-dotenv>=1.0.0",
    "ta>=0.10.0",
    "tavily-python>=0.3.0",
    "requests>=2.31.0",       
    "beautifulsoup4>=4.12.3",  
    "streamlit>=1.33.0",  
    "plotly>=5.20.0",  
    "fpdf2>=2.7.7", 
    "markdown2>=2.4.10",    
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

